moved spectrum analyzer message to piksi.yaml
