+++ javascript/sbp/piksi.js
@@ -694,6 +694,48 @@ MsgNetworkStateResp.prototype.fieldSpec.push(['tx_bytes', 'writeUInt32LE', 4]);
 MsgNetworkStateResp.prototype.fieldSpec.push(['interface_name', 'string', 16]);
 MsgNetworkStateResp.prototype.fieldSpec.push(['flags', 'writeUInt32LE', 4]);
 
+/**
+ * SBP class for message MSG_SPECAN (0x0050).
+ *
+ * Spectrum analyzer packet.
+ *
+ * Fields in the SBP payload (`sbp.payload`):
+ * @field t GPSTime Receiver time of this observation
+ * @field freq_ref number (float, 4 bytes) Reference frequency of this packet
+ * @field freq_step number (float, 4 bytes) Frequency step of points in this packet
+ * @field amplitude_ref number (float, 4 bytes) Reference amplitude of this packet
+ * @field amplitude_unit number (float, 4 bytes) Amplitude unit value of points in this packet
+ * @field amplitude_value array Amplitude values (in the above units) of points in this packet
+ *
+ * @param sbp An SBP object with a payload to be decoded.
+ */
+var MsgSpecan = function (sbp, fields) {
+  SBP.call(this, sbp);
+  this.messageType = "MSG_SPECAN";
+  this.fields = (fields || this.parser.parse(sbp.payload));
+
+  return this;
+};
+MsgSpecan.prototype = Object.create(SBP.prototype);
+MsgSpecan.prototype.messageType = "MSG_SPECAN";
+MsgSpecan.prototype.msg_type = 0x0050;
+MsgSpecan.prototype.constructor = MsgSpecan;
+MsgSpecan.prototype.parser = new Parser()
+  .endianess('little')
+  .nest('t', { type: GPSTime.prototype.parser })
+  .floatle('freq_ref')
+  .floatle('freq_step')
+  .floatle('amplitude_ref')
+  .floatle('amplitude_unit')
+  .array('amplitude_value', { type: 'uint8', readUntil: 'eof' });
+MsgSpecan.prototype.fieldSpec = [];
+MsgSpecan.prototype.fieldSpec.push(['t', GPSTime.prototype.fieldSpec]);
+MsgSpecan.prototype.fieldSpec.push(['freq_ref', 'writeFloatLE', 4]);
+MsgSpecan.prototype.fieldSpec.push(['freq_step', 'writeFloatLE', 4]);
+MsgSpecan.prototype.fieldSpec.push(['amplitude_ref', 'writeFloatLE', 4]);
+MsgSpecan.prototype.fieldSpec.push(['amplitude_unit', 'writeFloatLE', 4]);
+MsgSpecan.prototype.fieldSpec.push(['amplitude_value', 'array', 'writeUInt8', function () { return 1; }, null]);
+
 module.exports = {
   0x0069: MsgAlmanac,
   MsgAlmanac: MsgAlmanac,
@@ -734,4 +776,6 @@ module.exports = {
   MsgNetworkStateReq: MsgNetworkStateReq,
   0x00BB: MsgNetworkStateResp,
   MsgNetworkStateResp: MsgNetworkStateResp,
+  0x0050: MsgSpecan,
+  MsgSpecan: MsgSpecan,
 }
