+++ spec/yaml/swiftnav/sbp/piksi.yaml
@@ -438,3 +438,35 @@ definitions:
                 desc: IFF_BROADCAST - broadcast address valid
             - 0:
                 desc: IFF_UP - interface is up
+
+ - MSG_SPECAN:
+    id: 0x0050
+    short_desc: Spectrum analyzer
+    desc: |
+        Spectrum analyzer packet.
+
+    fields:
+        - t:
+            type: GPSTime
+            desc: Receiver time of this observation
+        - freq_ref:
+            type: float
+            desc: |
+                Reference frequency of this packet
+        - freq_step:
+            type: float
+            desc: |
+                Frequency step of points in this packet
+        - amplitude_ref:
+            type: float
+            desc: |
+                Reference amplitude of this packet
+        - amplitude_unit:
+            type: float
+            desc: |
+                Amplitude unit value of points in this packet
+        - amplitude_value:
+            type: array
+            fill: u8
+            desc: |
+                Amplitude values (in the above units) of points in this packet
