--- spec/yaml/swiftnav/sbp/specan.yaml
@@ -1,53 +1,0 @@
-# Copyright (C) 2017 Swift Navigation Inc.
-# Contact: Fergus Noble <fergus@swiftnav.com>
-#
-# This source is subject to the license found in the file 'LICENSE' which must
-# be be distributed together with this source. All other rights reserved.
-#
-# THIS CODE AND INFORMATION IS PROVIDED "AS IS" WITHOUT WARRANTY OF ANY KIND,
-# EITHER EXPRESSED OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE IMPLIED
-# WARRANTIES OF MERCHANTABILITY AND/OR FITNESS FOR A PARTICULAR PURPOSE.
-
-
-package: swiftnav.sbp.specan
-description: Spectrum analyzer and related interference messages
-render_source: True
-stable: True
-include:
-  - types.yaml
-  - gnss.yaml
-
-
-definitions:
-
- - MSG_SPECAN:
-    id: 0x0050
-    short_desc: Spectrum analyzer
-    desc: |
-        Spectrum analyzer packet.
-
-    fields:
-        - t:
-            type: GPSTime
-            desc: Receiver time of this observation
-        - freq_ref:
-            type: float
-            desc: |
-                Reference frequency of this packet
-        - freq_step:
-            type: float
-            desc: |
-                Frequency step of points in this packet
-        - amplitude_ref:
-            type: float
-            desc: |
-                Reference amplitude of this packet
-        - amplitude_unit:
-            type: float
-            desc: |
-                Amplitude unit value of points in this packet
-        - amplitude_value:
-            type: array
-            fill: u8
-            desc: |
-                Amplitude values (in the above units) of points in this packet
